% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{Casaux1998}
\alias{Casaux1998}
\title{Casaux1998}
\format{
A data frames containing diet data for two populations (Potter Cove & Harmony Point) of Harpagifer antarcticus from the South Shetland Islands.
\itemize{
  \item Record: The record, in this case the name for the two populations, Harpagifer_antarcticus_PotterCove and Harpagifer_antarcticus_HarmonyPoint for populations from Potter Cover and Harmony Point respectively.
  \item Prey: Name of the prey item consumes.
  \item PercentOccurrence: Data for the percent occurrence of the prey consumed.
  \item PercentNumber: Data for the percent number of the prey consumed.
  \item PercentWeight: Data for the percent weight of the prey consumed.
  }
}
\usage{
Casaux1998
}
\description{
Raw data from Casaux, 1998. Prey is listed in percent frequency, percent number, and percent mass consumed.
}
\details{
Data from Casaux, 1998 for the diet of Harpagifer antarcticus from two localities in the South Shetland Islands of Antarctica.
}
\references{
{
\itemize{
  \item Casaux R. 1998. The contrasting diet of Harpagifer antarcticus (Notothenioidei, Harpagiferidae) at two localities of the South Shetland Islands, Antarctica. Polar Biology 19:283-285.
  }
 }
}
\keyword{datasets}
