% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertFishbaseDiet.R
\name{ConvertFishbaseDiet}
\alias{ConvertFishbaseDiet}
\title{Converts FishBase/SealifBase volumetric diet data obtained from the diet function into a usable format for TrophicLevelR}
\usage{
ConvertFishbaseDiet(FishBaseDiet, ExcludeStage)
}
\arguments{
\item{FishBaseDiet}{a data frame produced by the rfishbase diet function}

\item{ExcludeStage}{a character, indicating which life stages to exclude. Must match stage names given by rfishbase (i.e. larvae, rec./juveniles, juv./adults, adults).}
}
\value{
a list of length two, with two data frames. One containing the re-formatted diet data and one containing the Taxonomy.
}
\description{
Converts FishBase/SealifBase volumetric diet data obtained from the diet function into a usable format for TrophicLevelR
}
\details{
This converts the data frame produced by rfishbase into a usable form for calculating trophic levels in TrophicLevelR.
}
\examples{
#Get rfishbase diet data for a few species
my.diets<-rfishbase::diet(c("Oreochromis niloticus","Salmo salar"))
#use the ConvertFishbaseDiet function to format it for TrophicLevelR and exclude recruits/juveniles
cleaned.diets<-ConvertFishbaseDiet(FishBaseDiet=my.diets, ExcludeStage=c("larvae","recruits/juv."))
}
