% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{CortesPreyVals}
\alias{CortesPreyVals}
\title{CortesPreyVals}
\format{
A data frame of of 10 rows and 6 columns
\itemize{
  \item FoodI: Food category I.
  \item FoodII: Food category II.
  \item FoodIII: Food category III.
  \item Stage: Life history stage of the prey item.
  \item TL: Trophic level of the prey item.
  \item SE: Standard error around trophic level estimate of the prey item.
}
}
\usage{
CortesPreyVals
}
\description{
A data frame containing prey items and their respective trophic levels for Chondrichthyes prey from Cortes, 1999
}
\references{
{
\itemize{
  \item Cortes E. 1999. Standardized diet compositions and trophic levels of sharks. ICES Journal of marine science 56:707-717.
  }
}
}
\keyword{datasets}
