% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{FishBasePreyVals}
\alias{FishBasePreyVals}
\alias{FishBasePreyVals}
\title{FishBasePreyVals}
\format{A data frame of of 384 rows and 6 columns
\itemize{
  \item FoodI: Food category I.
  \item FoodII: Food category II.
  \item FoodIII: Food category III.
  \item Stage: Life history stage of the prey item.
  \item TL: Trophic level of the prey item.
  \item SE: Standard error around trophic level estimate of the prey item.
}}
\usage{
FishBasePreyVals

FishBasePreyVals
}
\description{
A data frame containing prey items and their respective trophic levels following  FishBase and
 TrophLab.

A data frame containing search terms for plant mitochondrial loci. Can be subset for loci of 
 interest. Columns are as follows and users should follow the column format if they wish to
 add search terms using the MergeSearchTerms function:
}
\seealso{
\code{\link{MergeSearchTerms}}

\code{\link{FoodTroph}}
}
\keyword{datasets}
