% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoodTroph.R
\name{FoodTroph}
\alias{FoodTroph}
\title{Calculates trophic level from food item data for species, populations, and individuals using a random subsampling routine.}
\usage{
FoodTroph(FoodItems, PreyValues, Taxonomy)
}
\arguments{
\item{FoodItems}{a data frame with rows as individuals and each row consisting of a prey item name.}

\item{PreyValues}{a data frame with rows as prey item names and columns containing the trophic level of the prey item and the standard error of that trophic item.}

\item{Taxonomy}{a data frame with the least inclusive level in the leftmost column progressing to more inclusive with columns to the right.Can be a single column.}
}
\value{
a list containing estimated trophic levels from food items at each taxonomic level provided by the user.
}
\description{
Calculates trophic level from food item data for species, populations, and individuals using a random subsampling routine.
}
\examples{
#Get some food item data from rfishbase
my.foods<-rfishbase::fooditems(c("Anomalops katoptron", "Grammatorcynus bilineatus"))
#convert FishBase data into data for trophic calculation using TrophicLevelR
converted.foods<-ConvertFishbaseFood(my.foods)
#Load Prey Values
data(TrophLabPrey)
#Calculate Trophic Levels
my.TL<-FoodTroph(FoodItems = converted.foods$FoodItems,PreyValues = TrophLabPrey, Taxonomy = converted.foods$Taxonomy)
}
