% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{Herichthys}
\alias{Herichthys}
\title{Herichthys}
\format{
A data frame of of 519 rows and 40 columns
}
\usage{
Herichthys
}
\description{
Raw supplementary data from Magalhaes et al., 2015 for the diets of Herichthys minckleyi used in a tutorial for the vignette.
}
\references{
{
\itemize{
  \item Magalhaes IS, Ornelas-Garcia CP, Leal-Cardin M, Ramirez T, and Barluenga M. 2015. Untangling the evolutionary history of a highly polymorphic species: introgressive hybridization and high genetic structure in the desert cichlid fish Herichtys minckleyi. Mol Ecol 24:4505-4520. 10.1111/mec.13316
  }
}
}
\keyword{datasets}
