% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{Horn1982}
\alias{Horn1982}
\title{Horn1982}
\format{A list of length 2 containing two data frames.
\itemize{
  \item Available A data frame with four rows and seven columns containing data on prey availability in two different years.
  \item Consumed A data frame with four rows and seven columns containing data on prey consumption by two species in two different years..
  }}
\usage{
Horn1982
}
\description{
Raw data from Horn, 1982. Prey consumption and availability are in percent weight. Not that as the values do not sum to 100, the relative abundance will be calculated.
}
\details{
Data from Horn, 1982 for the diet of two species of Stichaeidae and prey availability over two years.
}
\references{
{
\itemize{
  \item Horn M, Murray S, and Edwards T. 1982. Dietary selectivity in the field and food preferences in the laboratory for two herbivorous fishes (Cebidichthys violaceus and Xiphister mucosus) from a temperate intertidal zone. Marine Biology 67:237-246.
  }
 }
}
\keyword{datasets}
