% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{SebastesStomachs}
\alias{SebastesStomachs}
\title{SebastesStomachs}
\format{
A data frames containing three columns with specimen identifier, specimen stomach weight, and specimen weight for eleven Sebastes flavidus.
\itemize{
  \item SpecimenSpecimen identifier. Follows the pattern of number, month, day, and year.
  \item StomachWeight: Weight of stomach contents in grams.
  \item PredatorWeight: Weight of the specimen in grams.
  }
}
\usage{
SebastesStomachs
}
\description{
Raw data from NOAA containing information on the stomach weight and specimen weight of eleven Sebastes flavidus caught in the Gulf of Alaska in 2011.
}
\details{
Stomach and specimen weights for eleven Sebastes flavidus from the Gulf of Alaska by NOAA in 2011 (available at: https://access.afsc.noaa.gov/REFM/REEM/WebDietData/DietDataIntro.php).
}
\references{
{
\itemize{
  \item Livingston PA, Aydin K, Buckley TW, Lang GM, Yang M-S, and Miller BS. 2017. Quantifying food web interactions in the North Pacific - a data-based approach. Environmental Biology of Fishes 100:443-470. 10.1007/s10641-017-0587-0.
  \item .
  }
 }
}
\keyword{datasets}
