% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DietTroph.R
\name{DietTroph}
\alias{DietTroph}
\title{Calculates trophic level from percentage based quantitative diet data (volumetric, weight, etc.) for species, populations, and individuals.}
\usage{
DietTroph(DietItems, PreyValues, Taxonomy, PreyClass = c("FoodI",
  "FoodII", "FoodIII", "Stage"), SumCheck = TRUE)
}
\arguments{
\item{DietItems}{A data frame with rows as individual entries and each row consisting of a prey 
classification and a corresponding diet percentages. The column names for prey classification of 
the diets should match those of PreyValues. The first column should contain the identifier of 
the individual and be named "Individual", The last column should contain the percent of the prey in the diet and be labelled "Percent".}

\item{PreyValues}{a data frame with rows as prey item names and columns containing the trophic 
level of the prey item and the standard error of that trophic item. The column names of 
PreyValues except for TL and SE should match those in DietItems.}

\item{Taxonomy}{a data frame with the least inclusive level progressing to more inclusive moving
towards the right.}

\item{PreyClass}{Column names of the PreyValues used for matching between DietItems and 
PreyValues, exclusive of TL and SE. Default is those of FishBase.}
}
\value{
a list length of the columns in taxonomy, each containing trophic level estimation at 
the respective taxonomic level.
}
\description{
Calculates trophic level from percentage based diet data following the routine described in TrophLab. 
While FishBase data obtained from rfishbase can be used, users can also upload their own data for
use in with function.
}
\examples{
#Get some food item data from rfishbase
library(rfishbase)
my.diets<-rfishbase::diet(c("Lutjanus apodus","Epinephelus itajara"))
#convert FishBase data into data for trophic calculation using TrophicLevelR
converted.diet<-ConvertFishbaseDiet(my.diets,ExcludeStage=c("larvae","recruits/juv."))
#Load Prey Values
data(FishBasePreyVals)
#Calculate Trophic Levels
my.TL<-DietTroph(DietItems = converted.diet$DietItems,PreyValues = FishBasePreyVals, Taxonomy = 
converted.diet$Taxonomy, PreyClass=c("FoodI","FoodII","FoodIII","Stage"))
}
\author{
Samuel Borstein
}
