% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DietTroph.R
\name{DietTroph}
\alias{DietTroph}
\title{Calculates trophic level from volumetric diet data for species, populations, and individuals.}
\usage{
DietTroph(Volumes, PreyValues, Taxonomy)
}
\arguments{
\item{PreyValues}{a data frame with rows as prey item names and columns containing the trophic level of the prey item and the standard error of that trophic item.}

\item{Taxonomy}{a data frame with the least inclusive level progressing to more inclusive moving towards the right.}

\item{volumes}{a data frame with rows as individuals and each row consisting of a prey item name and a corresponding volumetric percentage.}
}
\value{
a list length of the columns in taxonomy, each containing trophic level estimation at each taxonomic level
}
\description{
Calculates trophic level from volumetric diet data. Volumes must follow the fishbase format, with the columns for 
Individual/species, FoodI, FoodII, FoodII, stage, and volume.Prey values for TrophLab and
}
\examples{
#' #Get some food item data from rfishbase
my.diet<-diet(c("Oreochromis niloticus", "Salmo salar"))
#convert FishBase data into data for trophic calculation using TrophicLevelR
converted.diet<-ConvertFishbaseDiet(my.diet)
#Load Prey Values
data(TrophLabPrey)
#Calculate Trophic Levels
my.TL<-DietTroph(Volumes = converted.diet$FoodItems,Taxonomy = converted.diet$Taxonomy, TrophLabPrey)
}
