% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electivity.R
\name{Electivity}
\alias{Electivity}
\title{Calculates a variety of electivity indices and foraging ratios}
\usage{
Electivity(Diet, Available, Indices = c("ForageRatio", "Ivlev",
  "Strauss", "JacobsQ", "JacobsD", "Chesson", "VanderploegScavia"),
  Depleting = FALSE)
}
\arguments{
\item{Diet}{Data frame with data corresponding to consumed resources found in the diet. See details for formatting.}

\item{Available}{Data frame with data corresponding to the available resources. See details for formatting.}

\item{Indices}{Character vector containing the names of the desired indices to calculate. See description for information on availabble indices.}

\item{Depleting}{Logical. If true, will assume the food source is depleting and perform calculations following Manly. Default is FALSE.}
}
\value{
List containing data frames for each electivity index selected.
}
\description{
Calculates forage ratio and a variety of popular electivity indices.
}
\details{
This function calculates the forage ratio and a variety of electivity indices. Included indices include Ivlev's (1961),
Strauss' (1979), Jacob's Q and D (1974), Chesson's (1983)(Which is similar to Manl'y Alpha (1974)), and Vanderploeg & Scavia (1979).
The function takes two
}
\references{
Chesson, J. 1983. The estimation and analysis of preference and its relatioship to foraging models. Ecology 64:1297-1304.
Ivlev, U. 1961. Experimental ecology of the feeding of fish. Yale University Press, New Haven.
Jacobs, J. 1974. Quantitative measurement of food selection. Oecologia 14:413-417.
Manly, B. 1974. A model for certain types of selection experiments. Biometrics 30:281-294.
Strauss, R. E. 1979. Reliability Estimates for Ivlev's Electivity Index, the Forage Ratio, and a Proposed Linear Index of Food Selection. Transactions of the American Fisheries Society 108:344-352.
Vanderploeg, H., and D. Scavia. 1979. Two electivity indices for feeding with special reference to zooplankton grazing. Journal of the Fisheries Board of Canada 36:362-365.
}
