% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{FishBasePreyVals}
\alias{FishBasePreyVals}
\title{FishBasePreyVals}
\format{
A data frame of of 384 rows and 6 columns
\itemize{
  \item FoodI: Food category I.
  \item FoodII: Food category II.
  \item FoodIII: Food category III.
  \item Stage: Life history stage of the prey item.
  \item TL: Trophic level of the prey item.
  \item SE: Standard error around trophic level estimate of the prey item.
}
}
\usage{
FishBasePreyVals
}
\description{
A data frame containing prey items and their respective trophic levels following FishBase and
 TrophLab.
}
\references{
{
\itemize{
  \item Froese R, and Pauly D. 2018. FishBase. http://www.fishbase.org/2018).
  \item Pauly D, Froese R, Sa-a P, Palomares M, Christensen V, and Rius J. 2000. TrophLab manual. ICLARM, Manila, Philippines. 
  }
}
}
\keyword{datasets}
