% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{Horn1982}
\alias{Horn1982}
\title{Horn1982}
\format{
A list of length 2 containing two data frames.
\itemize{
  \item Available A data frame with 2 rows and 18 columns containing data on percent weight relative abundance of macroalgae prey availability in two different years.
  \item Consumed A data frame with four rows and 19 columns containing data on percent weight relative abundance of macroalgae prey consumption by two species in two different years.
  }
}
\usage{
Horn1982
}
\description{
Raw data from Horn, 1982. Prey consumption and availability are relative abundance by weight of macroalgae species. Note that values do not sum to 1 as dataset only includes macroalgae and not other prey items. See Table 4 and Table 4 in Horn et al., 1982.
}
\details{
Data from Horn, 1982 for the diet of two species of Stichaeidae and prey availability over two years.
}
\references{
{
\itemize{
  \item Horn M, Murray S, and Edwards T. 1982. Dietary selectivity in the field and food preferences in the laboratory for two herbivorous fishes (Cebidichthys violaceus and Xiphister mucosus) from a temperate intertidal zone. Marine Biology 67:237-246.
  }
 }
}
\keyword{datasets}
